+++ lib/sidekiq/lock.rb
@@ -1,9 +1,18 @@
-require "sidekiq/lock/version"
-require "sidekiq/lock/worker"
 require "sidekiq/lock/middleware"
 require "sidekiq/lock/redis_lock"
+require "sidekiq/lock/version"
+require "sidekiq/lock/worker"
 
 module Sidekiq
+
+  def self.lock_method
+    @lock_method || :lock
+  end
+
+  def self.lock_method=(method)
+    @lock_method = method
+  end
+
   module Lock
     THREAD_KEY = :sidekiq_lock
   end
@@ -14,3 +23,4 @@ Sidekiq.configure_server do |config|
     chain.add Sidekiq::Lock::Middleware
   end
 end
+