+++ lib/sidekiq/lock/worker.rb
@@ -2,8 +2,10 @@ module Sidekiq
   module Lock
     module Worker
 
-      def lock
-        Thread.current[Sidekiq::Lock::THREAD_KEY]
+      def self.included(base)
+        base.send(:define_method, Sidekiq.lock_method) do
+          Thread.current[Sidekiq::Lock::THREAD_KEY]
+        end
       end
 
     end